requirements.txt
python-telegram-bot>=20.0
httpx>=0.28.0
python-dotenv>=0.19.0
pytz>=2021.3
ollama>=0.6.0
